backend/requirements.txt
Django>=4.2
celery
django-celery-beat
pymongo
dnspython
gunicorn
pyinaturalist
python-dateutil
requests
django-environ
django-cors-headers
shapely
PyJWT
bcrypt
djangorestframework
djangorestframework-simplejwt